fix directory
